requirements.txt
pdfplumber>=0.10.0
psycopg2-binary>=2.9.9
openai>=1.30.0
python-dotenv>=1.0.0
